LayoutTests/accessibility/mac/aria-pressed-button-attributes.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="button1" role="button" aria-pressed="true">Foo</div>

<div id="button2" role="button" aria-pressed="false">Bar</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests role=button with aria-pressed will create an AXCheckbox:AXToggle on Mac with the right AXValue");

    if (window.accessibilityController) {

        var button1 = accessibilityController.accessibleElementById("button1");
        debug("Button1: " + button1.role);
        debug("Button1: " + button1.subrole);
        debug("Button1: " + button1.stringValue);

        var button2 = accessibilityController.accessibleElementById("button2");
        debug("Button2: " + button2.role);
        debug("Button2: " + button2.subrole);
        debug("Button2: " + button2.stringValue);
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
